+++ src/routes/products.router.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import ProductManagerMongo from '../dao/ProductManagerMongo.js';
-import { ProductModel } from '../models/product.model.js';
+
 
 const router = Router();
 const productManager = new ProductManagerMongo();
@@ -13,32 +13,6 @@ router.get('/', async (req, res) => {
     res.json(result);
   } catch (error) {
     res.status(500).json({ status: 'error', message: 'Error al obtener productos' });
-  }
-});
-
-//GET RENDER HBS
-router.get('/products', async (req, res) => {
-  try {
-    const page = parseInt(req.query.page) || 1;
-    const limit = 9;
-
-    const result = await ProductModel.paginate({}, {
-      page,
-      limit,
-      lean: true
-    });
-
-    res.render('products', {
-      products: result.docs,
-      page: result.page,
-      hasPrevPage: result.hasPrevPage,
-      hasNextPage: result.hasNextPage,
-      prevLink: result.hasPrevPage ? `/products?page=${result.prevPage}` : null,
-      nextLink: result.hasNextPage ? `/products?page=${result.nextPage}` : null
-    });
-  } catch (error) {
-    console.error('Error al cargar productos paginados:', error.message);
-    res.status(500).send('Error al cargar productos');
   }
 });
