+++ src/views/products.handlebars
@@ -9,7 +9,18 @@
 <body class="bg-gray-100 text-gray-900">
 
   <div class="container mx-auto px-4 py-8">
+
     <h1 class="text-3xl font-bold mb-6 text-center">Productos</h1>
+
+    <div class="flex justify-end mb-2">
+      <a class="group relative inline-block overflow-hidden border border-indigo-600 px-4 py-2 focus:ring-3 focus:outline-hidden" href="/carts">
+        <span class="absolute inset-y-0 left-0 w-[2px] bg-indigo-600 transition-all group-hover:w-full"></span>
+        <span class="relative text-sm font-medium text-indigo-600 transition-colors group-hover:text-white">
+          🛒 Ver carrito
+        </span>
+      </a>
+    </div>
+
 
     <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
       {{#each products}}
@@ -43,5 +54,3 @@
 
 </body>
 </html>
-
-
